+++ hooks/api.ts
@@ -1,4 +1,4 @@
-import { client, KyricsSWRResponse } from 'lib/api';
+import { client, clientWithoutToken, KyricsSWRResponse } from 'lib/api';
 import { useRouter } from 'next/router';
 import useSWR, { SWRConfiguration } from 'swr';
 import { ISongData, User } from 'types';
@@ -14,12 +14,10 @@ export const useGetUser = () => {
   return data?.data?.data;
 };
 
-export const useGetSongData = () => {
-  const router = useRouter();
-  const {
-    query: { id },
-  } = router;
-  const { data } = useSWR<{ data: { data: ISongData } }>(`/song/${id}`, client.get);
+export const useGetSongData = (id: string | string[] | undefined, user?: User | undefined) => {
+  const isToken = user ? client : clientWithoutToken;
+
+  const { data } = useSWR<{ data: { data: ISongData } }>(`/song/${id}`, isToken.get);
 
   return data?.data?.data;
 };
